Slightly changed architecture, fixed bugs
I tested the Trainer class, and noticed there were alot of bugs, still.
Now it seems to be working correctly. Changed the computation of the loss
to actually compute the mean of an array, such that there can be no error
of 'off by one' when dividing by the number of elements. The impact in terms
of memory and compute time should be negligible.
Note that logging of gradient is not possible with the current structure, as the process_batch happens
before gradient computation. I might want to rethink that.