Added lr cooldown, and debug param for steplogging
Added and tester lr coolodown (to a reasonable amount). Also added a
'debug' parameter into train_steps. If activated, will do logging
with '.detach()' instead of '.item()', to see if there is an impact in
performance. Will do beta testing, and if it works and is faster,
will change it to be default.